Enhancement: Adds ability to define custom docker volume path
In some pipeline environments (notably, Bitbucket Pipelines), it is
not possible to invoke a lambda due to the local volume being
hard-coded to `/tmp/data`, which is not guaranteed to be a writable
location.

I've added the ability to set an environment variable called
`MOTO_DOCKER_LOCAL_VOLUME_PATH`, which will be used in place of
`/tmp/data` when creating the local volume at lambda invocation time.

This is an optional environment variable; not setting it will result
in the original value of `/tmp/data` being used.